+++ addon/popup.js
@@ -6,12 +6,23 @@ import {setupLinks} from "./links.js";
 let h = React.createElement;
 
 {
-  parent.postMessage({insextInitRequest: true}, "*");
+  parent.postMessage({
+    insextInitRequest: true,
+    iFrameLocalStorage: {
+      popupArrowOrientation: localStorage.getItem("popupArrowOrientation"),
+      popupArrowPosition: JSON.parse(localStorage.getItem("popupArrowPosition")),
+      scrollOnFlowBuilder: JSON.parse(localStorage.getItem("scrollOnFlowBuilder"))
+    }
+  }, "*");
   addEventListener("message", function initResponseHandler(e) {
-    if (e.source == parent && e.data.insextInitResponse) {
-      removeEventListener("message", initResponseHandler);
-      init(e.data);
-      initLinks(e.data);
+    if (e.source == parent) {
+      if (e.data.insextInitResponse) {
+        //removeEventListener("message", initResponseHandler);
+        init(e.data);
+        initLinks(e.data);
+      } else if (e.data.updateLocalStorage) {
+        localStorage.setItem(e.data.key, e.data.value);
+      }
     }
   });
 }
@@ -204,7 +215,7 @@ class App extends React.PureComponent {
     let clientId = localStorage.getItem(sfHost + "_clientId");
     let hostArg = new URLSearchParams();
     hostArg.set("host", sfHost);
-    let linkInNewTab = localStorage.getItem("openLinksInNewTab");
+    let linkInNewTab = JSON.parse(localStorage.getItem("openLinksInNewTab"));
     let linkTarget = inDevConsole || linkInNewTab ? "_blank" : "_top";
     let browser = navigator.userAgent.includes("Chrome") ? "chrome" : "moz";
     return (
@@ -239,7 +250,7 @@ class App extends React.PureComponent {
               h("a", {ref: "limitsBtn", href: limitsHref, target: linkTarget, className: "page-button slds-button slds-button_neutral"}, h("span", {}, "Org ", h("u", {}, "L"), "imits"))
             ),
           ),
-          h("div", {className: "slds-p-vertical_x-small slds-p-horizontal_x-small"},
+          h("div", {className: "slds-p-vertical_x-small slds-p-horizontal_x-small slds-border_bottom"},
             // Advanded features should be put below this line, and the layout adjusted so they are below the fold
             h("div", {className: "slds-m-bottom_xx-small"},
               h("a", {ref: "metaRetrieveBtn", href: "metadata-retrieve.html?" + hostArg, target: linkTarget, className: "page-button slds-button slds-button_neutral"}, h("span", {}, h("u", {}, "D"), "ownload Metadata"))
@@ -279,6 +290,11 @@ class App extends React.PureComponent {
                   className: "page-button slds-button slds-button_neutral"
                 },
                 h("span", {}, "Setup ", h("u", {}, "H"), "ome")),
+            ),
+          ),
+          h("div", {className: "slds-p-vertical_x-small slds-p-horizontal_x-small"},
+            h("div", {className: "slds-m-bottom_xx-small"},
+              h("a", {ref: "options", href: "options.html?" + hostArg, target: linkTarget, className: "page-button slds-button slds-button_neutral"}, h("span", {}, "Options"))
             ),
           )
         ),
@@ -1874,7 +1890,7 @@ function sfLocaleKeyToCountryCode(localeKey) {
 }
 
 function getLinkTarget(e) {
-  if (localStorage.getItem("openLinksInNewTab") == "true" || (e.ctrlKey || e.metaKey)){
+  if (JSON.parse(localStorage.getItem("openLinksInNewTab")) || (e.ctrlKey || e.metaKey)) {
     return "_blank";
   } else {
     return "_top";